removed unused files from the codebase
