+++ apps/pwabuilder/src/script/pages/app-home.ts
@@ -15,10 +15,9 @@ import { isValidURL } from '../utils/url';
 
 import '../components/content-header';
 import '../components/loading-button';
-import '../components/dropdown-menu';
 import '../components/app-sidebar';
 import '../components/companies-packaged';
-import '../components/resource-hub-new';
+import '../components/resource-hub';
 import '../components/success-stories';
 import '../components/community-hub';
 
@@ -558,7 +557,7 @@ export class AppHome extends LitElement {
           </div>
         </div>
         <companies-packaged></companies-packaged>
-        <resource-hub-new></resource-hub-new>
+        <resource-hub></resource-hub>
         <success-stories></success-stories>
         <community-hub></community-hub>
       </main>
